refactor: extract inner classes from EventSourceManager
